Set explicit dtype to np.float32 in initializers to fix TypeError

+++ VishwamAI/scripts/model_architecture.py
@@ -15,19 +15,19 @@ class VishwamAIModel(hk.Module):
         self.tokenizer = keras_nlp.tokenizers.SentencePieceTokenizer(proto=tf.io.gfile.GFile(config.VOCAB_FILE, "rb").read(), sequence_length=1024, dtype="int32")
         self.transformer = hk.transform(
             lambda x: hk.Sequential([
-                hk.Embed(vocab_size=20000, embed_dim=128, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
+                hk.Embed(vocab_size=20000, embed_dim=128, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=np.float32)),
                 lambda x: self.attention(x, x, x),  # Keep inputs as integers for embedding
-                hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
-                hk.Linear(256, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
+                hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=np.float32)),
+                hk.Linear(256, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=np.float32))
             ])(x),
             apply_rng=True
         )
         self.attention = hk.MultiHeadAttention(
             num_heads=8,
             key_size=32,
-            w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")
+            w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=np.float32)
         )
-        self.dense = hk.Linear(3, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
+        self.dense = hk.Linear(3, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=np.float32))
 
         # Define expert networks for Mixture of Experts (MoE) architecture
         self.num_experts = 1  # Reduced number of experts to 1
